requirements.txt
Django==4.2.7
DjangoRESTframework==3.14.0
django-filter==23.4
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
